Split DevicesController: extract clone_device
Move clone_device into a sibling mutations_clone.py as a free
function taking controller as the first arg. The controller
keeps a thin bound-method delegate so WS dispatch and tests
calling controller.clone_device(...) still resolve unchanged.

Drops _rewrite_required_yaml_leaf, friendly_name_slugify,
generate_api_encryption_key, and rewrite_api_encryption_key
imports from controller.py (only used by the moved code).
configuration_stem, rewrite_name_or_substitution,
get_device_metadata, and set_device_metadata stay since other
still-on-controller methods (rename_device,
edit_friendly_name, _resolve_api_key_via_esphome_config) also
use them.

The test_clone_device_handles_filesystem_race_on_target_filename
test patches the open() builtin via the module namespace to
simulate a FileExistsError on the commit closure; re-pointed
at mutations_clone.open since the open() call moved with the
body.

Pure refactor; behaviour preserved. Drops ~167 lines from
controllers/devices/controller.py (now 1311).
